AVRO-4294: [c++] Validate enum/union index at decode; signed remaining subtraction
- decodeUnionIndex/decodeEnum now go through a shared decodeIndex() helper that
  rejects a negative decoded long (which would wrap to a huge size_t) and a value
  beyond size_t (which could truncate into a spuriously in-range branch on builds
  where size_t is narrower than int64_t) before casting. The concrete bound
  against the enum/union size is still enforced by the caller.
- MemoryInputStream2::remainingBytes() subtracts in int64_t so a violated
  invariant (curLen_ > size_) cannot underflow an unsigned subtraction into a
  huge positive remaining count.

+++ lang/c++/impl/Stream.cc
@@ -132,7 +132,10 @@ public:
     }
 
     int64_t remainingBytes() const final {
-        return static_cast<int64_t>(size_ - curLen_);
+        // Subtract in int64_t: if an invariant were ever violated (curLen_ >
+        // size_), doing the subtraction in unsigned size_t would underflow to a
+        // huge value and weaken the available-bytes guard.
+        return static_cast<int64_t>(size_) - static_cast<int64_t>(curLen_);
     }
 };
